♻️ refactor: refactor *.page.tsx to *.tsx
