+++ src/app/chat/layout.pc.tsx
