+++ src/app/chat/mobile/page.tsx
