+++ src/app/welcome/page.tsx
